MFC r281169, r293724, r293796, r294029, r294041, r294058
MFC r281169 (by andrew):
Make global variabled only used in this file static

MFC r294058:
Make common boot file_loadraw name parameter const

MFC r294041:
Remove unused reg param from fdt_fixup_memory

MFC r293724:
Enable warnings in EFI boot code

MFC r293796:
Fix typo in libefi.c

MFC r294029:
Only build EFI components on supported compilers

Sponsored by:	Multiplay


Former-commit-id: ecc7d12fac522a1591a979c692ff40d5002ed22d

+++ sys/boot/efi/include/efierr.h
@@ -46,7 +46,7 @@ name|EFI_ERROR_CODE
 parameter_list|(
 name|a
 parameter_list|)
-value|(a& ~EFI_ERROR_MASK)
+value|(unsigned long)(a& ~EFI_ERROR_MASK)
 end_define
 
 begin_define
